Share ruleset logic between Cli and debug mode
